Adding project report. Final images and test runs.

+++ src/meanStdImgs.m
@@ -1,6 +1,6 @@
 function [ feats ] = meanStdImgs( imgs )
     feats = [ 
         cellfun(@(x) mean(x(:)), imgs), ...
-        cellfun(@(x) std(x(:)), imgs)
+        cellfun(@(x) std(double(x(:))), imgs)
     ];
 end
